+++ examples/microprofile/oidc/src/main/java/io/helidon/examples/microprofile/security/oidc/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2020 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.security.examples.idcs;
-
-import java.util.Collections;
-import java.util.Set;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
 /**
- * TODO javadoc.
+ * An OIDC (Open ID Connect) example.
  */
-@ApplicationPath("/jersey")
-public class JerseyApplication extends Application {
-    @Override
-    public Set<Class<?>> getClasses() {
-        return Collections.singleton(JerseyResource.class);
-    }
-}
+package io.helidon.examples.microprofile.security.oidc;