+++ microprofile/security/src/main/java/io/helidon/microprofile/security/SecurityMpService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -75,7 +75,7 @@ public class SecurityMpService implements MpService {
         Config webServerConfig = config.get("security.web-server");
         if (webServerConfig.exists() && webServerConfig.get("enabled").asBoolean().orElse(true)) {
             context.serverRoutingBuilder()
-                    .register(WebSecurity.create(security, config));
+                    .register(WebSecurity.create(security, config.get("security")));
         }
 
         // for later use, e.g. for outbound security